+++ bindings/nodejs/lib/types/block/core/basic.ts
@@ -4,7 +4,7 @@
 import { Payload, PayloadDiscriminator } from '../payload';
 import { Type } from 'class-transformer';
 import { StrongParents, WeakParents, ShallowLikeParents } from '../parents';
-import { Block } from '.';
+import { Block } from './block';
 
 /**
  * Basic Block layout.
